backend/requirements.txt
# Django + Extensions
Django==2.0.5
djangorestframework==3.8.2
django-crispy-forms==1.7.2
django-request-logging==0.6.5

# Testing
Faker==0.8.8
mixer==6.0.1
coverage==4.5.1

# API-Docs
coreapi==2.3.3
Markdown==2.6.11
Pygments==2.2.0
